requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-google-genai
langchain-huggingface
sentence-transformers
chromadb
pypdf
requests
beautifulsoup4

